Fixed incorrect column type for food journal entry quantity; added boolean flags for include food item in recipe or in food journal.

+++ DatabaseAccess/Entities/FoodJournalEntry.cs
@@ -7,7 +7,7 @@ namespace SnackAndTrack.DatabaseAccess.Entities {
         public virtual required Guid Id { get; set; }
         public virtual required DateOnly Date { get; set; }
         public virtual required TimeOnly? Time { get; set; }
-        public virtual required Int16 Quantity { get; set; }
+        public virtual required Single Quantity { get; set; }
         public virtual required Unit Unit { get; set; }
         public virtual required FoodItem FoodItem { get; set; }
     }
